adds oh my zsh and node installs

+++ setup_node.zsh
@@ -2,12 +2,16 @@
 
 echo "\n<<< setting up node >>>\n"
 
-# if exists node; then
-#   echo "node exists"
-# else
+if exists node; then
+  echo "node exists"
+else
   # Dig out nvm
   . "$NVM_DIR/nvm.sh"
-  # Install tools
   nvm install 12
+fi
+
+if exists yarn; then
+  echo "yarn exists"
+else
   npm i -g yarn
-# fi
+fi
